fix ci multiple status reports

+++ github_bot.rb
@@ -136,13 +136,13 @@ class GithubBot
       request[:name] = "Nervos Integration"
     end
 
-    check = installation_client.get("/repos/nervosnetwork/ckb/commits/#{payload['check_run']['head_sha']}/check-runs", accept: accept)['check_runs'].find do |check|
-      check['name'] == request[:name] && check['app']['name'] == 'Nervos Bot' && check['head_sha'] == request[:head_sha]
+    installation_client.get("/repos/nervosnetwork/ckb/commits/#{payload['check_run']['head_sha']}/check-runs", accept: accept)['check_runs'].each do |check|
+      if check['name'] == request[:name] && check['app']['name'] == 'Nervos Bot' && check['head_sha'] == request[:head_sha] then
+        request.delete(:head_sha)
+        installation_client.patch("/repos/nervosnetwork/ckb/check-runs/#{check['id']}", request)
+      end
     end
-    if check
-      request.delete(:head_sha)
-      installation_client.patch("/repos/nervosnetwork/ckb/check-runs/#{check['id']}", request)
-    else
+    if request[:head_sha]
       installation_client.post('/repos/nervosnetwork/ckb/check-runs', request)
     end
   end
